Use correct mixin
